scripts/matter_provisioning/requirements.txt
optigatrust 
Pillow 
qrcode 
brother_ql 
pyusb